2048-sqrt: Add "√", "×2", and "×4" tiles. Win at 32768.
A derivative of baiqiang's "Advanced 2048"
http://baiqiang.github.io/2048-advanced/
but with square-root tiles added as well.

Don't allow the player's score to go negative due to square-rooting
a tile whose value was originally built up on the cheap using the
"×2" and "×4" multiplier tiles. We do this by remembering the total
number of points earned during the construction of this tile, and
when its value decreases, we compute a new value ("as if" it had been
constructed fairly) and make you lose that many points. The number of
points earned during construction is stored in the tile's "accumulatedScore"
member.

Also, since the ×2 and ×4 tiles make it pretty easy to get to 2048,
make the win condition 32768 (as in "Advanced 2048").

FYI, yes, combining two fourth-root tiles ought to give you a "16th-root"
tile instead of merely an "eighth-root" tile (in the sense that taking
the fourth root of the fourth root of X is equivalent to taking the
16th root of X); but if I did that, you'd never be able to clear the
tile; whereas it's plausible to clear an eighth-root tile by combining
it with 256.  (×4 tiles do combine in the mathematically correct way.)